:sparkles: arrange installations

+++ init.sh
@@ -1,2 +1,10 @@
+if [ ! $(which cargo) ]; then
+    # install Rust and tool chains
+    curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
+fi
+
 ln -sf $(pwd)/cargo-atcoder.toml ~/Library/Preferences/
-brew install binutils
+
+if [ ! -e /usr/local/opt/binutils/bin/strip ]; then
+    brew install binutils
+fi
